leave bytes out of this!!!!

+++ ipmifanctl/logic.py
@@ -31,7 +31,7 @@ class Logic:
         speed_range = (self.hi_speed - self.lo_speed)
         interpolant = (self._peak_reading - self.lo_temp) / temp_range
         t = pow(interpolant, self.curve_power) # Power-curved interpolant
-        self._value = self.lo_speed + speed_range * t
+        self._value = int(self.lo_speed + speed_range * t)
 
     @property
     def value(self):
